+++ docs/changelog.md
@@ -7,6 +7,8 @@
   - 新增 `weekday` 和 `isoweekday` 方法
   - 补充公历2101年日期的干支、节气信息
   - 修正无节气的日期格式化的bug
+- `calendars.festivals` 模块
+  - 修正 `SolarSchema` 二月倒数序号形式解析的bug
 - 新增 `borax.structures.proxy` 模块
 - `trim_iterable` 新增前缀、后缀字符串